sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentlens-observe"
version = "0.1.0"
description = "Debug AI agents visually — self-hosted, open-source observability SDK"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "AgentLens Contributors"}]
keywords = ["ai", "agents", "observability", "tracing", "debugging", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Debuggers",
]
dependencies = ["httpx>=0.27"]

[project.optional-dependencies]
langchain = ["langchain-core>=0.2"]
crewai = ["crewai>=0.51"]
autogen = ["autogen-agentchat>=0.4"]
llamaindex = ["llama-index-core>=0.11"]
google-adk = ["google-adk>=0.3"]
otel = ["opentelemetry-api>=1.20", "opentelemetry-sdk>=1.20"]
all = ["agentlens-observe[langchain,crewai,autogen,llamaindex,google-adk,otel]"]

[tool.hatch.build.targets.wheel]
packages = ["agentlens"]

[project.urls]
Homepage = "https://github.com/tranhoangtu-it/agentlens"
Repository = "https://github.com/tranhoangtu-it/agentlens"
Issues = "https://github.com/tranhoangtu-it/agentlens/issues"
